+++ src/app/routes/Index.tsx
@@ -21,9 +21,9 @@ import Details from "../ui/pages/details/Details";
 import AI from "../ui/pages/AI/AI";
 import Website from "../ui/pages/website/Website";
 import WebWelcome from "../ui/pages/website/webwelcome/WebWelcome";
-import KnowledgeCityDetails from "../ui/pages/knowledgecitydetailspage/KnowledgeCityDetails";
-import Cashbasket from "../ui/pages/cashbasket/Cashbasket";
-import Dome from "../ui/pages/dome/Dome";
+import KnowledgeCityDetails from "../ui/pages/animations/knowledgecitydetailspage/KnowledgeCityDetails";
+import Cashbasket from "../ui/pages/animations/cashbasket/Cashbasket";
+import Dome from "../ui/pages/animations/dome/Dome";
 import WebOverview from "../ui/pages/website/weboverview/WebOverview";
 import Project from "../ui/pages/website/project/Project";
 import WebFoarm from "../ui/pages/website/webfoarm/WebFoarm";
