+++ src/app/ui/pages/animations/cashbasket/Cashbasket.css
