+++ src/app/ui/pages/website/webwelcome/WebWelcome.tsx
@@ -74,51 +74,6 @@ const WebWelcome: React.FunctionComponent = () => {
             </div>
           </div>
         </section>
-        {/* TEST SECTION */}
-        <div>
-          <Button
-            bgColor={"#fbcc34"}
-            mTop={0}
-            mBottom={0}
-            mLeft={0}
-            mRight={0}
-            fWeight={800}
-            bRadiusColor="#fbcc34"
-            title="Overview of your previous"
-            color={"#071d69"}
-            onClickButton={() => {
-              navigate("/weboverview");
-            }}
-          />
-          <Button
-            bgColor={"#fbcc34"}
-            mTop={0}
-            mBottom={0}
-            mLeft={0}
-            mRight={0}
-            fWeight={800}
-            bRadiusColor="#fbcc34"
-            title="View Your Projects"
-            color={"#071d69"}
-            onClickButton={() => {
-              navigate("/project");
-            }}
-          />
-          <Button
-            bgColor={"#fbcc34"}
-            mTop={0}
-            mBottom={0}
-            mLeft={0}
-            mRight={0}
-            fWeight={800}
-            bRadiusColor="#fbcc34"
-            title="View selection"
-            color={"#071d69"}
-            onClickButton={() => {
-              navigate("/webfoarm");
-            }}
-          />
-        </div>
       </main>
     </>
   );
